+++ src/app/modules/review/review.services.ts
@@ -2,6 +2,12 @@ import prisma from "../../utils/prisma";
 import { IReview } from "./review.interface";
 
 const createReviewDb = async (payload: IReview) => {
+  const productExists = await prisma.product.findUnique({
+    where: { id: payload.productId },
+  });
+  if (!productExists) {
+    throw new Error("Invalid productId: Product not found");
+  }
   const result = await prisma.review.create({
     data: payload,
   });
@@ -30,6 +36,13 @@ const fetchReviewDb = async (productId?: string) => {
 };
 
 const fetchSingleReviewDB = async (productId: string) => {
+  const productExists = await prisma.review.findUnique({
+    where: { id: productId },
+  });
+  if (!productExists) {
+    throw new Error("Invalid productId: Product not found");
+  }
+
   return await prisma.review.findUnique({
     where: { id: productId },
     include: { user: true },
@@ -37,6 +50,13 @@ const fetchSingleReviewDB = async (productId: string) => {
 };
 
 const updateReviewDb = async (payload: Partial<IReview>, reviewId: string) => {
+  const reviewExists = await prisma.review.findUnique({
+    where: { id: reviewId },
+  });
+  if (!reviewExists) {
+    throw new Error("Invalid productId: Product not found");
+  }
+
   const result = await prisma.review.update({
     where: { id: reviewId },
     data: payload,
@@ -45,6 +65,12 @@ const updateReviewDb = async (payload: Partial<IReview>, reviewId: string) => {
 };
 
 const deleteReviewDb = async (reviewId: string) => {
+  const reviewExists = await prisma.review.findUnique({
+    where: { id: reviewId },
+  });
+  if (!reviewExists) {
+    throw new Error("Invalid productId: Product not found");
+  }
   const result = await prisma.review.delete({
     where: { id: reviewId },
   });
