+++ src/app/modules/shippingAddress/shippingAddress.services.ts
@@ -1,47 +1,51 @@
 import prisma from "../../utils/prisma";
 import { IShippingAddress } from "./shippingAddress.interface";
 
+// Create new shipping address
+const createDB = async (data: IShippingAddress) => {
+  return await prisma.shippingAddress.create({
+    data: data,
+  });
+};
+
+// Get all shipping addresses of a user
+const getByUserDB = async (userId: string) => {
+  return await prisma.shippingAddress.findMany({
+    where: { userId },
+    orderBy: { createdAt: "desc" },
+  });
+};
+
+// Update address
+const updateDB = async (id: string, data: any) => {
+  const exitingId = await prisma.shippingAddress.findUnique({
+    where: { id: id },
+  });
+  if (!exitingId) {
+    throw new Error("Params id is Invalid ");
+  }
+  return await prisma.shippingAddress.update({
+    where: { id: id },
+    data,
+  });
+};
+
+// Delete address
+const deleteDB = async (id: string) => {
+  const exitingId = await prisma.shippingAddress.findUnique({
+    where: { id: id },
+  });
+  if (!exitingId) {
+    throw new Error("Params Id  is Invalid ");
+  }
+  return await prisma.shippingAddress.delete({
+    where: { id: id },
+  });
+};
 
 export const ShippingAddressService = {
-  // Create new shipping address
-  create: async (data: IShippingAddress) => {
-    const address = await prisma.shippingAddress.create({
-      data: data
-    });
-    return address;
-  },
-
-  // Get all shipping addresses of a user
-  getByUser: async (userId: string) => {
-    const addresses = await prisma.shippingAddress.findMany({
-      where: { userId },
-      orderBy: { createdAt: "desc" },
-    });
-    return addresses;
-  },
-
-  // Get single address
-  getById: async (id: string) => {
-    const address = await prisma.shippingAddress.findUnique({
-      where: { id },
-    });
-    return address;
-  },
-
-  // Update address
-  update: async (id: string, data: any) => {
-    const updated = await prisma.shippingAddress.update({
-      where: { id },
-      data,
-    });
-    return updated;
-  },
-
-  // Delete address
-  delete: async (id: string) => {
-    const deleted = await prisma.shippingAddress.delete({
-      where: { id },
-    });
-    return deleted;
-  },
+  getByUserDB,
+  createDB,
+  deleteDB,
+  updateDB,
 };
